+++ src/flipper/component.js
@@ -449,7 +449,7 @@ Component.prototype = {
             }
         }
 
-        return templateEngine.renderView(viewId, data, options);
+        return templateEngine.renderView(viewId, data, options, this);
     },
 
     /* created / attached cycle methods */
@@ -494,6 +494,10 @@ Component.prototype = {
         } else if (element.hasAttribute('model-id')) {
             modelId = element.getAttribute('model-id');
             result = Flipper.dataCenter.getSpace(modelId);
+        } else if (element.hasAttribute('model-key')) {
+            modelId = '';
+            utils.log('"model-key" is a test feature, do not use');
+            result = window[element.getAttribute('model-key')];
         }
 
         return Promise.resolve(result).then(function(model) {